feat: add bitwise operations

+++ types/boxed.go
@@ -20,7 +20,7 @@ func IsBoxable(v int64) bool {
 }
 
 func BoxI32(v int32) Boxed {
-	return box(uint64(v), KindI32)
+	return box(uint64(uint32(v)), KindI32)
 }
 
 func BoxI64(v int64) Boxed {
